requirements.txt
certifi==2025.11.12
massive==2.0.2
polygon-api-client==1.16.3
python-dotenv==1.2.1
urllib3==2.5.0
websockets==15.0.1
